requirements.txt
# Core Dependencies
zipline-reloaded>=3.0.0
empyrical-reloaded>=0.5.9
exchange-calendars>=4.5.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Metrics & Analysis
empyrical>=0.5.5
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Data Sources
yfinance>=0.2.36
pandas-datareader>=0.10.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Development
ipykernel>=6.25.0
jupyter>=1.0.0
pytest>=7.4.0

